rename package, add a simplest unit test
