+++ app/build.gradle
@@ -20,6 +20,7 @@ repositories {
 dependencies {
     // This dependency is used by the application.
     implementation 'com.google.guava:guava:30.1.1-jre'
+    testImplementation 'junit:junit:4.12'
 }
 
 
@@ -35,5 +36,5 @@ dependencies {
 }
 application {
     // Define the main class for the application.
-    mainClass = 'com.Demo.RatpackDemo.App'
+    mainClassName = 'com.corposense.ocr.demo.App'
 }
