+++ app/src/main/java/com/corposense/ocr/demo/App.java
@@ -1,11 +1,14 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package com.Demo.RatpackDemo;
+package com.corposense.ocr.demo;
 
 import ratpack.server.RatpackServer;
 
 public class App {
+
+    final public static int CONSTANT_VALUE = 1;
+
     public static void main(String... args) throws Exception {
         RatpackServer.start(server -> server
                 .handlers(chain -> chain
